Add files via upload
Replace cim.py in the python_cim. This version doesn't require the WMI db files to be the exact names. As long as the WMI db files start with the correct name and the extension contains the correct extension, it will use those files. This is gear to for doing mass triage.