Add files via upload
tarprocess -- used to process like tar files. For example, linux triage scripts collect files and put them into a tar file. The scripts will extract the files, hash them, and save a copy to a directory. The file/directory permissions and UID/GID is recreated from the information in the tar file. 

Files are save as {filename}_{hash}